More info about the last commit: UI, improved site block, Pomodoro habit task.
Many fixes and improvement to site blocking and site block overlay of Pomodoro. added a lot of images and improved User interface, new colors and Pomodoro character. A Pomodoro habit is created automatically in Habitica, with an option in the settings to automatically score [+] or [-] of the habit when Pomodoro is done successfully or fails accordingly.